Extract the list to a common component
